:truck: Split models into separate modules
Organizing by DigiD and eHerkenning, similarly to the SAML models.
